Move thunk code generation into its own, smarter, function.
In the general case, we want a thunk any time we pass something as an argument
or let-bind it to a variable, so we call thunk expr. However, certain
functions deal with unlifted types and some expressions are already thunks,
and we don't want to create new thunks for any of those. Moving thunking to
its own function, we can deal with this in one place.