Generate thunks if and only if appropriate.
This commit is perhaps overly aggressive in that it may result in raw values
getting passed around where thunks should be, where said raw values are in
HNF. This requires the eval operation to cope with both thunks and raw values,
which might be undesirable if we were going for raw speed.

In any event, care must be taken not to break E() in the runtime library for
raw values.