Create and match on constructors properly, using their tags for ID.
Objects are created by calling the RTS function D with the constructor's tag
and arity and 0 or more of the constructor's arguments. The function will
return either a data object, represented by an array, or a function that
creates a data object when saturated.

This means that any of the following is fine:
D(0, 3, a, b, c)
D(0, 3)(a)(b)(c)
D(0, 3, a)(b, c)
etc.