Fold up nested abstractions and applications into multiarg abs/app.
Not only is it inefficient and spacious to do f(x)(y)(z) or
function(x) {
  return function(y) {
    ...
  }
}
...but it's also hideous and makes the output harder to debug.

This commit does away with such constructs as far as possible, packing the
above examples up info f(x,y,z) and function(x,y) {...} respectively.

Note that this needs runtime support to work in the general case; partial
application obviously breaks with this change, so now we have to do one of:
a) route all calls through an Apply operation that takes care of the gritty
   details;
b) insert code to handle this at each call site, or;
c) do a) but attempt to bypass the Apply middle-man when we are absolutely
   positive that the number of arguments matches up with the function's
   arity.

For a start, we should do a), and upgrade that to c) in the future. b) would
certainly be faster, but since small code has priority over fast code, it's
a no-go.