Implement all double primops except DoubleDecode_2IntOp.
DoubleDecode_2IntOp will be hard or impossible to implement; it requires
extracting the sign, exponent and mantissa from a double. It'd be easy if we
could just peek at the bits behind the numbers but as JS doesn't allow that,
we'd have to be clever.

First thought on the matter is that we're looking at a factorization problem,
which is NP. Just rewriting any parts of the standard library that depend on
it is probably a good idea in general.