Remove primop generation, runtime library and symbol table from TODOS.
While neither primop generation nor runtime library are what you'd call
"done," they now evolve along with the rest of everything, rather than
standing apart as separate unimplemented work items.

The symbol table, it seems, will not be necessary; rather, we'll replace the
external names (which are the only names we need to do something about) still
in use at link time and leave it at that; the Core names serve just as well
in the JS AST as in the Core source.