Clean up JS names, retain both external and unique names for external IDs.
There was quite a lot of cruft to do with naming that it turns out we didn't
really need, so I just got rid of it. Also, both external and unique names
are now preserved for external calls; when the linking step rolls around,
we're going to want to use those unique names.