Force saturation of nullary data constructors on creation.
Data constructors need to be fully saturated before we can do any matching on
them. This wasn't a problem for n>0:ary constructors as it's impossible to use
them in Core without at least one application of them, but nullary
constructors stayed as closures in the generated JS, which of course ruined
any code that needed to use them.