Eliminate var -> var assignments at the end of unconditional branches.
Whenever we encounter a case expression with a single alternative, we know
two things: it's only there to evaluate something, and the lone alternative
will always be executed. Normally, we assign the result of a case alternative
to a variable, as any code coming after the case expression must know where
to find its result. However, if we only have one alternative, we might as well
tell them where the result of that branch is, rather than stuff it into a
previously known location.