Generate a name for GHC.Prim.coercionToken#, but keep it undefined for now.
There seems to be no information whatsoever available about what this
coercionToken actually is, but nothing seems to ever examine it. If we keep
it as undefined, things will explode loudly whenever someone DOES look at it,
so we can at least detect what's going on and fix it then.