Fix foo `triggers` (bar `triggers` async baz)
Before, we only lifted Async signals outside Trigger signals when compiling,
which meant that we didn't take care of the case described above.
Now we check for Async immediately when applying triggers instead, ensuring
that Async is always lifted out of Trigger, regardless of how much nesting
is going on.