Optimize most tail recursion into loops, even without --opt-tce.
"Most" meaning any tail recursive call which does not occur in a mutually
recursive function. This restriction is in place because of an extremely
obscure bug that occurs when using tail loops in certain mutually recursive
functions. (integerTimes, for example, gives the wrong result for certain
arguments when the result is >2^31 and this restriction is not used.)