Always get properties and styles of elements as strings.
Before, we assumed that every element style or property was a string; now we
force it to be so. In the long term, a type safe library ought to be build on
top of this, but getting and setting arbitrary properties can't really be done
in a type safe manner.

This commit fixes issue 7.