Haste is now bootstrapped by haste-boot; refuses to run if unbooted.
In order to get rid of the magic in buildall.sh, the build process is now
split into two parts: the build, consisting of running cabal install for
first Fursuit, then Haste; and the bootstrapping, which consists of fetching
the base libraries, fetching the Closure compiler and generating intermediate
code for the Fursuit and Haste libraries, and is invoked by simply running
the haste-boot binary.

After boot, .haste/booted contains the compiler's boot version, as given in
HasteReboot.hs. This version should be increased whenever a change needing a
re-bootstrap of Haste is introduced. Should hastec detect that its internal
boot version is greater than the one in .haste/booted, it will refuse to run
unless invoked with the --unbooted command line argument.

If Haste is already properly booted, haste-boot will do nothing, unless
invoked with --force.