Allow Haste to do partial reboots, to avoid redownloading base needlessly.
Often, the Haste.* modules change while the code generator remains the same.
In such cases the Haste standard library needs to be reinstalled but the
base package is still fine, so it would be wasteful to redownload everything.

This commit lets Haste do and require either a partial reboot, for stdlib
changes, or a full reboot, when the code generator changes.