New functions for appending DOM nodes; old ones deprecated.
The old functions had a convention of putting the child to append as their
first arguments, not the parent as one would expect. Apart from being
confusing, this also does not play very nice with currying: appending several
child nodes to a single parent is quite a common operation whereas appending
a single child to multiple parents doesn't even make sense.

The new functions follow the more sensible convention of putting the parent
node first, and the argument to append last.