Fix hypothetical bug with closure-creating tail loops.
If a closure-creating tail loop were to return 0, '', null, or some other
value which JavaScript considers equal to null, Haste would interpret this
as a signal to keep running the loop instead of returning it. This commit
introduces a sentinel object as the "keep going" signal, which JavaScript
only ever considers equal to itself.